Restructured strategy module
Split strategy module into the main module, which contains SudokuInfo
and the Strategy definition, the solver-module which contains the
definition of the StrategicSolver and StrategicBacktrackingSolver, and
the impls-module which contains the actual strategies.
